Вынес верстку из контроллера в отдельный файл
